requirements.txt
apache-airflow[gcp]==2.5.0
google-cloud-dataproc==5.0.0
google-cloud-storage==2.7.0
google-cloud-pubsub==2.14.1
